requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docformatter==1.3.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
untokenize==0.1.1
urllib3==1.25.8
wrapt==1.11.2
